requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
cycler==0.10.0
idna==3.2
joblib==1.1.0
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.3
pandas-ta==0.3.14b0
Pillow==8.3.2
plotly==5.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
six==1.16.0
tenacity==8.0.1
threadpoolctl==3.0.0
urllib3==1.26.7
yfinance==0.1.63
